streaming/requirements.txt
requests==2.31.0
boto3==1.34.0
pandas==2.3.1
kafka-python==2.0.2
confluent-kafka==2.3.0
pyspark==3.5.0
delta-spark==3.0.0 